+++ src/pages/sign-in/index.js
@@ -1,0 +1,1 @@
+export { default as SignIn } from "./SignIn";
