+++ src/widgets/sign-in-form/index.js
@@ -1,0 +1,1 @@
+export { default as SignInForm } from "./ui/SignInForm";
